app/src/androidTest/java/br/com/jluiz20/baseproject/presentation/main/MainActivityTest.java
package br.com.jluiz20.baseproject.presentation.main;

import android.app.Instrumentation;
import android.support.test.InstrumentationRegistry;
import android.support.test.rule.ActivityTestRule;
import android.support.test.runner.AndroidJUnit4;

import org.junit.Before;
import org.junit.Rule;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.ArgumentCaptor;
import org.mockito.junit.MockitoJUnit;
import org.mockito.junit.MockitoRule;

import javax.inject.Inject;

import br.com.jluiz20.baseproject.MockBaseProjectApplication;
import br.com.jluiz20.baseproject.R;
import br.com.jluiz20.baseproject.di.TestComponent;

import static android.support.test.espresso.Espresso.onView;
import static android.support.test.espresso.assertion.ViewAssertions.matches;
import static android.support.test.espresso.matcher.ViewMatchers.withId;
import static android.support.test.espresso.matcher.ViewMatchers.withText;
import static org.mockito.Mockito.clearInvocations;
import static org.mockito.Mockito.verify;

/**
 * Tests for {@link MainActivity}
 *
 * @author João Luiz Vieira <vieira.jluiz@gmail.com>.
 */
@RunWith(AndroidJUnit4.class)
public class MainActivityTest {

    @Rule
    public ActivityTestRule<MainActivity> activityRule = new ActivityTestRule<>(
            MainActivity.class,
            true,     // initialTouchMode
            true); // launchActivity. False so we can customize the intent per test method

    @Rule
    public MockitoRule mockitoRule = MockitoJUnit.rule();

    @Inject
    MainActivityContract.Presenter presenter;

    private ArgumentCaptor<MainActivityContract.View> viewArgumentCaptor;

    @Before
    public void setUp() {
        Instrumentation instrumentation = InstrumentationRegistry.getInstrumentation();
        MockBaseProjectApplication app = (MockBaseProjectApplication) instrumentation.getTargetContext().getApplicationContext();
        TestComponent component = (TestComponent) app.component();
        component.inject(this);
    }

    @Test
    public void test_AppVersion() {
        activityRule.getActivity().runOnUiThread(() -> {
            viewArgumentCaptor = ArgumentCaptor.forClass(MainActivityContract.View.class);
            verify(presenter).onViewResume(viewArgumentCaptor.capture());
        });
        activityRule.getActivity().runOnUiThread(() -> viewArgumentCaptor.getValue().showAppVersion("1.5.4"));

        onView(withId(R.id.app_version))
                .check(matches(withText("1.5.4")));

        clearInvocations(presenter);
    }


}
